automater/workflows/serializers.py
from rest_framework import serializers
from .models import Workflow, ExecutionLog

class WorkflowSerializer(serializers.ModelSerializer):
    class Meta:
        model = Workflow
        fields = "__all__"

class ExecutionLogSerializer(serializers.ModelSerializer):
    class Meta:
        model = ExecutionLog
        fields = "__all__"
