automater/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9.9
requests>=2.31.0
